+++ NEWS.md
@@ -1,3 +1,6 @@
++ Added Anderson-Darling test `ds.anderson.darling.test`.
++ `ds.exposure_histogram` now checks the number of samples and if over 5000 uses `ds.anderson.darling.test` instead of `ds.shapiro.test`.
+
 # dsExposomeClient 2.0.2
 
 + Added datasources to function calls inside `ds.exposure_histogram`